requirements.txt
asgiref==3.7.2
Django==4.2.1
django-countries==7.5.1
djangorestframework==3.14.0
pytz==2024.1
sqlparse==0.4.4
tzdata==2024.1
